Migration bugs. - Fixed foreign keys referred to 'users' table because the table uses big integer increment field since it's migrated to Laravel 7.

+++ database/migrations/2018_09_24_121006_create_table_account.php
@@ -17,7 +17,7 @@ class CreateTableAccount extends Migration
             $table->increments('id');
             $table->string('name', 100)->nullable(false);
             $table->integer('type_id')->unsigned()->nullable(false);
-            $table->integer('user_id')->unsigned()->nullable(false);
+            $table->bigInteger('user_id') ->unsigned()->nullable(false);
             $table->decimal('balance', 10, 2)->default('0.00');
 
             $table->foreign('type_id', 'fk_account_type_id')
